create skeleton pages
