+++ client/pages/chat.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export default () => <div>Page for teacher-student chat</div>
